requirements.txt
python-dotenv
pandas
numpy
google-genai
dtaidistance
